requirements.txt
psutil>=5.8.0
distro>=1.6.0
boto3>=1.20.0
botocore>=1.23.0
certifi>=2021.10.8
setuptools<81
PyYAML>=6.0
configparser>=5.0.0
